Fix book creation and editing, Add two column layout

+++ src/main/java/com/toastedbits/bookish/controllers/BookController.java
@@ -6,9 +6,14 @@ import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.toastedbits.bookish.domain.Book;
+import com.toastedbits.bookish.domain.Category;
+import com.toastedbits.bookish.exceptions.ResourceNotFoundException;
 import com.toastedbits.bookish.services.BookService;
+import com.toastedbits.bookish.services.CategoryService;
 
 @Controller
 @RequestMapping("/book")
@@ -17,26 +22,54 @@ public class BookController {
 	private BookService bookService;
 	
 	@Autowired
-	GalleryHelper helper;
+	private CategoryService catService;
 	
 	@RequestMapping(value="/{id}", method=RequestMethod.GET)
 	public String get(ModelMap model, @PathVariable("id") Long id) {
-		model.put("book", bookService.getById(id));
-		helper.setModelAttributes(model);
+		Book book = bookService.getById(id);
+		if(book == null) {
+			throw new ResourceNotFoundException();
+		}
+		model.put("book", book);
+		model.put("categories", catService.getAll());
 		return "book";
 	}
 	
 	@RequestMapping(value="/{id}", method=RequestMethod.DELETE)
 	public String delete(ModelMap model, @PathVariable("id") Long id) {
+		if(bookService.getById(id) == null) {
+			throw new ResourceNotFoundException();
+		}
 		bookService.deleteById(id);
 		return "redirect:/books";
 	}
 	
 	@RequestMapping(method=RequestMethod.PUT, value="/{id}")
 	public String put(ModelMap model, @PathVariable("id") Long id, Book book) {
+		if(id == null || bookService.getById(id) == null) {
+			throw new ResourceNotFoundException();
+		}
+		
+		book.setCategory(catService.fetch(book));
 		bookService.updateById(id, book);
-		model.put("book", book);
-		helper.setModelAttributes(model);
-		return "book";
+		return "redirect:/book/" + id;
+	}
+	
+	@RequestMapping(method=RequestMethod.GET, value="/{id}/content")
+	public @ResponseBody String getContent(@PathVariable("id") Long id) {
+		Book book = bookService.getById(id);
+		if(book == null || book.getContent() == null) {
+			throw new ResourceNotFoundException();
+		}
+		return book.getContent();
+	}
+	
+	@RequestMapping(method=RequestMethod.GET, value="/{id}/summary")
+	public @ResponseBody String getSummary(@PathVariable("id") Long id) {
+		Book book = bookService.getById(id);
+		if(book == null || book.getSummary() == null) {
+			throw new ResourceNotFoundException();
+		}
+		return book.getSummary();
 	}
 }
